code refactor

+++ minsu3d/model/module/backbone.py
@@ -37,7 +37,7 @@ class Backbone(pl.LightningModule):
         output_dict = {}
         x = ME.SparseTensor(features=voxel_features, coordinates=voxel_coordinates)
         unet_out = self.unet(x)
-        output_dict["point_features"] = unet_out.features[v2p_map.long()]
+        output_dict["point_features"] = unet_out.features[v2p_map]
         output_dict["semantic_scores"] = self.semantic_branch(output_dict["point_features"])
         output_dict["point_offsets"] = self.offset_branch(output_dict["point_features"])
         return output_dict
